new scripts
confirm takes two arguments - the first is the prompt, what dmenu wants you to confirm, and the second is the command to run if the answer is yes. pomodoro is a simple shell script pomodoro timer. linkhandler asks you what program you want to open a link with, where the link is the argument you pass it.